refactor(app): decouple ipc broadcast as will be used from database
